Removed Test JS
Removed test JS as it did not have any functionality.
Added the "Required" spans and "Error" spans in HTML. The purpose of the required span is to alert the user if they did not fill in the input.
The purpose of the error span is to alert the users if the data was filled incorrectly. e.g if the user puts letters of the alphabet in the phone number input